Some u_long -> unsigned int changes.
These are valid because the invention and major uses of u_long in this
codebase predate the 64-bit transition - it may look like we're
narrowing size from 64 to 32 bits, but what we're actually doing is
replacing an earlier hack for guaranteeing 32-bit size by people
who cut their teeth on 16-bit ints.

We might actually get a little optimization out of this, but the
real reason is to get rid of the obfuscatory typedef.