A contemplated 'snapshot' feature in the build recipe is unnecessary.
There was a TODO item that read:

* Add 'snapshot' feature to dump config status into a JSON file for collecting
  build + platform information

This was missing the obvious; there already is a parseable representation
of that state, it's called the config.h file.  On the Don't Repeat Yourself
principle, if we want JSON or anything else we should make it by simple
postprocessing rather than adding needless complexity to the build recipe.

Added: a PLATFORM_FULL feature, the string from Python's platform.platform(),
to capture more information.