Address GitLab issue  Issue #13: DNS lookup broken in obscure cases
This is "A hack good-enough solution would be to make the array bigger."
Before this change the default array sizes were absurdly small, as though
RAM still cost the way it did in 1990 or whenever.

Also added: a note in devel/TODO that this whole mess ought to be replaced
by something like c_ares.