ntpd: Fix another place good data being thrown away.
ntpd assumed that no refclock measurement could be more precise
that sys_precision.  Measurements smaller than sys_precision were
replaced with sys_precision, even if the measure was zero or averaged.

This assumption is not valid when KPPS is used.  The KPPS timestamps
can be much more precise than sys_precision.  Sometimes more than 60%
of the good and precise data was being discarded.

Removing this bad code can be shown to decrease RMS Time Jitter
by a factor of 4 on a Raspberry Pi.

This problem became more obvious with Linux kernel 2.4.24 on RasPi.
This kernel degrades sys-precision from -20 to -19.  That moved the
floor of RMS Time Jitter from 1 micro second to 2.