ntpviz: Autorange a bit differently in clip mode.
Normally, if most of the data is around 1 few micro sec, but max
data is ten seconds, the autorange unit is set to one second.

In clip mode this can truncate almost all the data.  So in clip mode
autorange to the 1% and 99% points.  Now 5 micro sec will show as
5.000 microsec, not 0.000 sec.