Make authentication code from libntp available in the Python extension.
This required breaking some adhesions. Notably (1) progname now lives in
msyslog.c, which is what uses it, and (2) the authkeys and msyslog stuff
can no longer call the ISC assert macros.

The latter could be fixed by relocating the ISC assertion stuff into
the list of sharables the Pyton extension is built with, but using
those macros there was probably not the best of ideas to begin with.

Python interfaces for the auth functions aren't written yet.