Zero-fill the unused part of the Mode 6 response packets.
This wasn't needed when the clients were all in C, for which the first
NUL is a string terminator.  But Python allows NULs in strings, which
means Python mode 6 clients actually see the trailing garbage.

There's another motivation as well.  I'm chasing a bug that inserts
binary junk in textual Mode 6 responses. Zeroing the packet buffer
before each send will help locate this.