Fix part of the binary-tumor bug in Mode 6 ctl_putdata()
The way data was copied before a flush was unnecessary and combined
with some still unknown bug to produce binary garbage in the *middle*
of tags and values.

The cost of removing this extra complication is that because data
items can no longer cross fragment boundaries multipacket responses
will be slightly less space-efficient.

Binary crap is still being emitted, but only on item boundaries.