Change pre-requisites to indicate Python 2 must be at 2.6.6 or later.
This was brought on by GitLab Issue #198.  Under Python 2.6.4 waf
throws an AttributeException while trying to do subprocess control.
This problem is not known to arise on Centos 6, which also runs 2.6.

I've set the low bar to 2.6.6 rather than 2.6.5 from a dim memory
of having seen this problem (and this fix) years ago.