Make shm report POSIX seconds since epoch, not calendar time.
There is both a philosophical and a practical point to this change.
The philosophical point is that the device driver logging should
report what it actually *knows*, not add a whole layer of
interpretation that is potebtially buggy.  The practical point is that
we are trying to narrow the cope of use of the homebrew vint64 type
preparatory to abolishing it.