Repair a problem in Mode 6 protocol handling.
The old C ntpq code was, as it turns out, broken.  It assumed that a nonce
issued by ntpd was good for 4 succeeding requests when, in fact, the server
would consider it invalid 16 seconds after the receipt time of the request
for the nonce.

The Python Mode 6 protocol code inherited this bug.  This commit fixes
the problem.