Remove the Jupiter/Zodiac driver.  It cannot have worked since August 1999.
The way we know this is that its timecode interpretation assumed cyclic GPS
time starting from the GPS Epoch 0 date at the beginning of 1980. Thus,
reported time will have warped back to 0 at the Epoch 1 turnover and
been wrong ever since.  That was a bit over 17 years ago at the time of
this commit - and Epoch 1 firmware in the device wouldn't fix it.