Address GitLab issue #316: "ntpd" dies unexpected unpeering a refclock
The root of this problem is the deprecation of referring to refclocks
by magic addresses.  A user tried unpeering a refclock by driver type name
and fatally confused the peer lookup logic.

No point in trying to fix this, since in the absence of a unit number
the type specification is ambiguous.  So just put in a crash guard and
log in a complaint.