refclock_trimble: require event polling for non-Thunderbolt receivers
Remove the ability to disable event polling because packets 8f-ad and 8f-0b
are needed for the rollover workaround (#461) but the Palisade and Acutime
receivers transmit only packet 8f-ad if their event inputs aren't used. Running
with event polling disabled is normally done for troubleshooting purposes, but
it's also used for some Palisades with firmware that doesn't allow event
inputs. A no-cost firmware upgrade is available which enables event inputs
for those receivers.