refclock_trimble: fix leap-scheduled detection, simplify polling
Testing of the leap detection logic with a Spirent GSS6100 GPS
signal generator revealed that the logic for Palisade and Acutime
receivers was incorrect. The leap-second-flag bits (packet 8f-ad
byte 19) are the same for both insertion and deletion, so assume
insertion.

To simplify polling: don't count packets, instead finish the poll
with the last packet in the sequence sent after an event or
auto-report.