refclock_trimble: fix for classic bug 1075
The simplest solution of waiting a few microseconds between setting
and clearing RTS causes problems with the Acutime 2000 and Praecis.
The Acutime 2000 may become unresponsive if the pulse time is
between 5us and 600ms, and the Praecis will occasionally produce
event timestamps that are incorrect by a few us when the pulse width
is between ~5us and 100us.
Extending the RTS set time to the entire polling interval works for
all supported receivers, but a sleep for at least 12ms is needed
between clearing and re-setting RTS for the receivers to re-arm
their event inputs.