Address GitLab issue #214: waf "host" part of the build doesn't use our CFLAGS
This change propagates compiler dialect switches to both the host and main
parts of the build.  Doing this with -std=c99 exposed a couple of minor
port bugs, which are also fixed.

This should fix at least one build issue under Solaris.

NOTE: While this change is not *known* to create any port problems with
unusual platforms, it's the kind of thing that very well could. Consider
bisecting here in case of port breakage.