Fixed scrambled data storage in SyncPacket()
Previously SyncPacket would take data and store it in self.data, which
would only ever be used by analyze. Then analyze() would parse self.data
leaving cruft behind in self.extension. flatten() encoded the header and
tacked on self.extension. The result is that a packet could not be encoded
and decoded losslessly.