Fixes most warnings from gcc 4.0.1.
Most of the warnings are related to the named formal parameters in the
propototypes for step_systime() and adj_systime(), which shadow the
globals settime and adjtime, respectively.  It's not clear why only
this particular GCC version complains, but since the other formal
parameters in those definitions are already unnamed, removing those
names is more consistent, anyway.

The other eliminated warnings relate to a couple of GCC pragmas which
are unsupported prior to 4.2.

The remaining warnings fall into two groups:

1) When building against the MacPorts Python 2.7, it gets a couple of
"shadowed global" warnings from Python.h.  Since this issue is
entirely within Python's code, there really isn't anything ntpsec can
do about it.

2) Although the warnings about unrecognized warning-disable pragmas in
tests/libntp/msyslog.c are now eliminated, the warnings they are
trying to suppress are still present.  This is only in the
msnprintfHangingPercent test, which is trying to test behavior in this
specific illegal case, so the warnings are completely expected when
not suppressed.  This an issue with GCC <4.2.  A cleaner fix for
this whole mess of warning disablement might be to move this test to
a separate source compiled without the warnings enabled in the first
place.

TESTED:
Still builds clean on OSX 10.6, 10.9 & 10.12, and Ubuntu 14.04.
Builds with only the abovementioned warnings on OSX 10.5.