Simplify passing configuration data to the peer-creation logic.
Before, everything passed tp newpeer was explicit arguments to it.
Now we pass in the struct peerctl control block created by config.
Where newpeer() is called by the protocol machine we assemble such a
block on the fly.

This change has two advantages:

(1) if we need to create new config items, adding them to the peer ctl
structure suffices to get them all the way into the newpeer()
logic. It's no longer required that we futz with the newpeer() calls
on the way down - though the passed-in values might need validation
or massaging un newpeer() itself.

(2) The protocol-machine calls are actually more readable this way,
because the code setting up the control blocks makes explicit semantic
information that was previously only implied by the order of arguments.

No logic changes in this commit.