Allow refclock sample processing to use 4-digit year timestamps.
This is actually a pretty big deal.  Before this change the refclock sample
processing simply ignored the year part of timestamps passed in from, e.g.,
GPS devices.  As a defensive adaptation to devices that shipped only 2 digit
timestamps this made sense - the century was, in effect, deduced from the
receive timestamp of the sample.

But this meant that interpretation of the sample was contingent on the
system clock being correct, and bad things could happen if it was
trashed or (especially) zeroed, returning it to 1970.  After that you
couldn't use a refclock reliably until you'd already achieved time
sync with a network peer.

With this change, an ntpd receiving proper 4-digit timestamps (e.g. from
the NMEA driver) no longer needs the system clock to be correct, and can
run without remote peers.  Recovery from a trashed system clock can happen.

A unit test verifying this capability is included in the patch.