Restructure the peer block to keep the peerctl structure intact.
Members that used to be copied ad hoc from the peer_ctl structure into
separate slots in a new peer structure are now, instead, copied into
a struct peer_ctl that retains their grouping.

This commit does no logic changes.  It is preparation for cleaning up the
logic in newpeer() that does the copies to do the block copy directly.
The ultimate goal is to eliminate friction when defining new struct peer_ctl
members in the configurator.