Fixes clock_gettime()/clock_settime() fallbacks.
This replaces the broken and incomplete OSX-specific clock_gettime()
fallback for a partially generic version based on gettimeofday().  The
fallback for CLOCK_MONOTONIC is still OSX-specific (but without the
former port leak bug).  It also adds generic fallback for
clock_settime() based on settimeofday().

Using gettimeofday()/settimeofday() is actually optimal on OSX, in
spite of what one might expect of clock_get_time() and
clock_set_time().

OSX builds get a couple of warnings on 10.6, and many on 10.5, but
these are unrelated to the change.

TESTED:
Builds and runs on OSX 10.5-10.12, including 10.5 PPC.