Adds clock_gettime/clock_settime fallback to ntpfrob.
Unfortunately, there doesn't seem to be an easy way to incorporate
clockwork.c into ntpfrob, due to a morass of dependencies, so this
just duplicates the code into ntpfrob.  Not ideal, but not too awful
in this context.

TESTED:
Ran ntpfrob on OSX 10.9.  The "ntpfrob -e" result is a little wonky,
but that code has multiple issues, anyway.