refclock_nmea: provisional fix for issue #62.
In the NMEA refclock, when PPS is lost, the peer jitter was left at the
low value calculated for PPS.  This jitter is now used for NMEA.  So
the NMEA time is wrongly thought to be PPS precise.  This caused the
selection algorithm to wrongly pick NMEA over other better sources.

The fix is to reset the jitter when PPS is lost.