Remove Magnavox driver.
None of these units have been manufactured since 1994, when the
manufacturer was bought by Leica Geosystems.  Original MX4200s shipped
invalid times after the GPS Era 1 rollover in 1999; Leica announced an
update PROM in 1997 but it was to cost $100 and is not even certain it
actually shipped. In 2017 these receivers can't even be found on eBay.