Fixes Python in-tree build setup.
This reworks a few things to make in-tree testing of Python code
work properly.  Note that all tweaks for this are now applied to
the build tree rather than the source tree, so the build-tree
copies of the programs should be used.  Specifically:

1) Moves the generated Python sources to the build tree, where they're
findable by the Python programs in the build tree.

2) Builds the C extension into the pylib directory, where it's needed.
This avoids the need to symlink it into pylib, and to be aware of the
different naming conventions used by different Python versions.

3) It copies the non-generated Python sources to the build tree, where
they're sometimes needed even when the .pyc/.pyo files are present
(e.g., by Python 3).

4) The symlinks for the 'ntp' package are now placed in the build tree
rather than the source tree.

TESTED:
Ran three client programs (ntpdig, ntpq, and ntpviz) sufficient to
cover all imports, as well as the four test programs in tests/pylib.
No import-related issues, though some programs have other issues with
Python 3.