Adds cleanup for incorrect Python library installs.
Because install and uninstall only apply to the locations known to the
current version of the script, libraries installed to the location
determined by the old buggy code could linger.  This could be
especially troublesome if PYTHONPATH had (inappropriately) been set up
to point to the incorrect location.  This adds code to get around this
by removing any libraries installed to the old location (when it
differs).  Since one might do an install without an uninstall, this
cleanup is applied to both.

TESTED:
Ran both install and uninstall, each after previously installing with
the old code.  Verified that the old libraries were removed, and that
the warning to remove the old directory appeared.