Removes all PYTHONPATH references.
All references to PYTHONPATH existed solely in relation to the
incorrect library install locations that may have been used prior to
the recent fix.  Since setting PYTHONPATH is generally undesirable, it
has now been removed from the error messages and from the
documentation.

Note that there were two import errors in ntpmon unrelated to ntpsec
itself, but which suggested the use of PYTHONPATH.  If this is still
an issue, something else needs to be fixed.

TESTED:
Ran all relevant programs with the ntp libraries inaccessible, and
verified proper error messages.