Fixes FixConfig to ignore existing PYTHONPATH.
Since the purpose of FixConfig is to avoid paths which only work when
specified by PYTHONPATH, any existing PYTHONPATH definition needs to
be avoided when checking a path.  This is done by removing PYTHONPATH
from the environment used to run the external Python that obtains the
sys.path setup.

TESTED:
Ran "configure" on Ubuntu 14.04, CentOS 7, and Fedora 25.  Verified
that the resulting PYTHONDIR is unchanged in the normal no-PYTHONPATH
case.  Verified that defining PYTHONPATH to match OLD_PYTHONDIR no
longer alters the result where it perviously did (CentOS and Fedora).