Fix ntploggps
I ported ntploggps to the new-style API and made it wait for a valid SKY
message.  This fixes the delivery of TDOP (which was otherwise always
zero for me) and the satellite count (which was often zero).  It also
waits for a valid TPV message, ensuring it can report the device
properly.

With this, my GPS reporting fully works on Ubuntu 16.04 with gpsd 3.15.