systemd: Cleanup ntp-wait.service
It seems more proper to do WantedBy=time-sync.target, and it achieves
the proper effect.  Then, WantedBy implies a reverse Before, so the
Before=time-sync.target is redundant.

I assume the original arrangement was copied from systemd-timesyncd.
That is more analogous to ntpd, not ntpwait.  In fact, systemd needs to
add a systemd-timesyncd-wait (which would be analogous to ntp-wait):
https://github.com/systemd/systemd/issues/5097