Fix compiler warnings on Solaris in ntpfrob/main.c
../../ntpfrob/main.c: In function 'main':
../../ntpfrob/main.c:40:2: warning: implicit declaration of function 'getopt' [-Wimplicit-function-declaration]
  while ((ch = getopt(argc, argv, "a:Ab:cejp:rh?")) != EOF) {
  ^

Commit "Reduce clutter from ./waf -v" caused definitions like __EXTENSIONS__
and _POSIX_C_SOURCE to now only be defined in config.h, instead of being in the
command line passed to the compiler.

Therefore we must include config.h before including system headers to get the
correct feature macros to be set so that we get POSIX versions of functions and
symbols.