Use PYTHONDIR and PYTHONARCHDIR explicitly
This removes wafhelpers/fix_python_config.py.  That had two purposes:
1) To avoid passing the PREFIX to get_python_lib().  This was done
   because the result is not always in the default sys.path.  However,
   this violates PREFIX.  For example, the default prefix of /usr/local
   would result in a PYTHONDIR under /usr/lib on systems without any
   /usr/local in sys.path.
2) Cleaning up files installed to both the new and old directories.

I have addressed #1 in a different way.  The PYTHONDIR and PYTHONARCHDIR
are written into the utilities at install time using waf's substitution.
Then, those directories are inserted into sys.path, *after* the entry
for the current working directory.  The original sys.path is restored
after importing the ntp.* modules, such that subsequent imports of other
modules use the normal sys.path.

I have chosen not to address #2.  I think the value of cleaning up after
old source installs is insufficient to justify the complexity and risk
of messing around with waf in the way wafhelpers/fix_python_config.py
did.  If others disagree, portions of that will need to be kept.  In
that case, note that the "old" and "new" directories need to be
switched, as we are back to passing the PREFIX to get_python_lib().

This fixes #414.