buildprep script should ensure compiler is installed
While trying to build NTPsec on a new laptop, I realised
that the CC compiler and friends are not being pulled in
explicitly.  They do get installed on Ubuntu by the
avahi dependency, but that is marked in the text as
optional.

If someone has an old SuSE, please check that install.