Continue if unable to contact a server
The ntpsweep utility will probably not do what you think
it does, as most NTP servers will refuse to be probed.

Instead of failing hard if unable to contact a server, we now
continue to the next server in the list.  This works only when
you are specifying a breadth of servers, if probing up the tree
recursively ( -p ), there is no next server.