Makes pythonize-header use waf's Python.
Aside from the nominal inconsistency of running Python-based
build helpers with a different version of Python than the one
running waf, this can actually fail when the system default Python
is too old for the helper code.  This change arranges to use the
waf Python for pythonize-header, similarly to what is already
being done for the tests.

TESTED:
With default Python set to 2.5, and 2.7 explicitly invoked to run waf,
verified that the build fails without this change and works with it.