Eliminates unnecessary symlink recreation.
This makes two changes to the symlink creation code in tests/wscript:

1) It avoids creating the symlink when it already exists and has the
correct content.

2) It creates it as a relative symlink rather than an absolute
symlink, so that it doesn't clash with the duplicate logic in
"afterparty".

Fixing this is a prerequisite to allowing installs from read-only
trees, though there is still a problem in that area from waf itself.

TESTED:
The correct relative symlink is now constructed prior to "afterparty",
and the tests pass.  Running "install" from a read-only mount no
longer crashes in the symlink code, though it still dies from waf's
attempt at some sort of database update.