Removes gratuitous reruns of pythonize-header.
The code which unconditionally removed the outputs of pythonize-header
was completely unnecessary, and not only caused unnecessary rebuilding
of those files, but also made it impossible to perform an install from
a read-only tree (after building read-write, of course).

Without this code, the targets are still rebuilt whenever the relevant
sources changed, and not rebuilt redundantly.

TESTED:
Verified that the targets are rebuilt when needed, and not rebuilt
when not needed.