Fix PPS documentation
The PPS documentation did not match the code.

The code is:
  if (typepps != NULL && fabs(sys_offset) < 0.4 &&
      (!typepps->is_pps_driver ||
      sys_prefer != NULL ||
      (typesystem == NULL && sys_minsane == 0))) {

Note that the sys_prefer and sys_minsane rules only apply to the driver
named pps, not to drivers with integrated PPS support.  PPS always needs
another source, but for integrated drivers, that source is the same
driver.

The documentation also said the PPS driver will be selected "if the PPS peer
is designated as a prefer peer", but that is not true; no such code exists.