Fix a bug in handling of the rate-limiting KOD.
The Classic code did the equivalent of this - not identical because
some structures gave been refactored:

	if (rbufp->pkt.ppoll > peer->cfg.minpoll)
		peer->cfg.minpoll = peer->ppoll;
	poll_update(peer, rbufp->pkt.ppoll);

During the Great Refactoring of the protocol machine, Daniel changed
it to this:

	if (peer->cfg.minpoll < 10) { peer->cfg.minpoll = 10; }
	poll_update(peer, 10);

Thw second line prevented the polling rate from being set arbitrarily
high by a malicious KOD from the server.  But the first line
introduced a bug - on receipt of any rate-limiter KOD the client would
get stuck at poll interval 10 and never recover.

This change reverts to allowing the KOD to both adjust the poll
interval and the minimum poll interval, but shortstops malicious
attempts to set them to unhelpfully high values.