Allow "prefer" on a "pps" peer to use it
This change allows "prefer" on a "pps" peer to indicate that the PPS
input should be used with any source.

The documentation, prior to Richard Lagger's recent cleanup, listed
this as an option, but it did not exist in the code:

"3. If there is a PPS driver and the system clock offset at this point
is less than 0.4 s, and if  there is a prefer peer among the survivors
or if  the PPS  peer is designated  as a prefer  peer, the  PPS driver
becomes the system peer"

Note the option of "if the PPS peer is designated as a prefer peer".

This adds that functionality, which is useful if you want to use the
PPS input with any source. The typical example here would be that you
want to use network sources to number the seconds. Without this
change, you have to mark one as a prefer peer, and the PPS is only
used if that particular peer is selected. This has a couple
problems. First, if that peer goes down, the PPS stops being
used. Second, marking a particular network peer as a prefer peer gives
it special status in the clock selection algorithm, which may not be
desirable.

This resolves GitLab issue #538.  It's a merge by hand of an MR
by Richard Laager.