Use BSD-4-Clause-UC
In cdb90975b71e46f9438ffa6c1449b615c7c882cd, the SPDX identifier was
changed from BSD-4-clause to BSD-2-clause because "The Regents of the
University of California have deprecated the 4-clause licence."

SPDX has a specific ID for this license, which "Captures the
retroactive deletion of the third (advertising) clause".

Note that, in any event, this is not a 2-clause BSD license, as only one
clause was removed from the 4-clause license, leaving 3 clauses.