Remove CI jobs for Debian "oldoldstable"
"oldoldstable" is two releases behind the current stable release.  It's not
supported by upstream so we should not build for it.

And the CI builds for "wheezy" broke today when "oldoldstable"'s repo was moved
from deb.debian.org to archive.debian.org as announced at
https://lists.debian.org/debian-devel-announce/2019/03/msg00006.html

The container images at https://hub.docker.com/_/debian and
https://github.com/debuerreotype/docker-debian-artifacts/blob/3e751c2c2f60037e9231ed94bbd1f95347af2c87/oldoldstable/slim/Dockerfile
have not yet caught up with the new repo location.