Fix ntpdate -s (syslog)
The ntpdate wrapper script was converting -s (for "log to syslog") to
ntpdig -p.  This is wrong, as ntpdig -p is for the number of samples and
requires a parameter.  The ntpdig man page says, "This version does not
log to syslog. Pipe standard output and standard error to logger(1) if
you want this behavior.

Signed-off-by: Richard Laager <rlaager@wiktel.com>