Add many formatting changes ...
The debug and standard HTML pages contain broken variable substitution that
render to things like "the {ntpfrob} program" instead of "the ntpfrob(8)
program" and "be reported by {ntpq}/{ntpmon}." instead of "be reported by
ntpq(1)/ntpmon(1)."

The ntpd-body includes simmilar but different errors like "the
/etc/{ntpconf} file" instead of "the /etc/ntp.conf file"

The access-commands and clock-options include problems with redefining
already placed anchors. Possible results are seeking to the wrong place in
the document when corresponding links are clicked.