Fix a comment
IEEE Std 1003.1-2017 says:
Since this volume of POSIX.1-2017 is aligned with the ISO C standard,
and since all functionality enabled by _POSIX_C_SOURCE set equal to
200809L is enabled by _XOPEN_SOURCE set equal to 700, there should be
no need to define _POSIX_C_SOURCE if _XOPEN_SOURCE is so defined.
Therefore, if _XOPEN_SOURCE is set equal to 700 and _POSIX_C_SOURCE is
set equal to 200809L, the behavior is the same as if only _XOPEN_SOURCE
is defined and set equal to 700. However, should _POSIX_C_SOURCE be set
to a value greater than 200809L, the behavior is unspecified.