Remove -M option, a noop, and correct documentation
In commit d4691e602d8906ce2eff8fd1638dfa3b65cee3dd
Hal dropped support for the SHA1 and MD5 lines in
ntp.keys, and the code.

The man page still mentioned that -M , which is documented
by NTP Classic as:

    Generate a new keys file containing 10 MD5 keys and 10 SHA keys.
    An MD5 key is a string of 20 random printable ASCII characters,i
    while a SHA key is a string of 40 random hex digits.

was the default, and that ntpkeygen operated _always_ in
MD5 mode.