Config: Fix error when miscasting pointer ...
There is a warning:
'cast from pointer to integer of different size [-Wpointer-to-int-cast]'
which breaks ./waf config when it reaches the function check for
CMAC_CTX_new if CFLAGS= "-Werror" is prepended. This seems to solve it.

Single casting as (long) is unacceptable.

Returning 'p ? 1 : 0' or '!!p' breaks the macOS runners;
they detecct res_init() which is not a good thing.

Thus the double cast (int)(long)