Rate limiting cleanup
The old rate limiting scheme was carefully tuned to work with
a single copy of the same code.  That doesn't work if you have
several clients behind a NAT box.

That chunk of code now uses a simple exponential decay (aka
leaky bucket) to keep track of recent traffic.  The time constant
is 20 seconds and the rate limit is 1 packet per second.  There is
no config option to change them.  (yet?)  The "score" is scaled
such that the units are packets/second.

That pair of parameters means that a client can start with a burst
of 20 packets but over the long term, but can only sustain 1 packet
per second.

ntpq/mru now displays the score and the number of packets
from that IP Address that were dropped.