Update license for NTS code to be BSD-2-Clause
This code was created by Hal Murray for the NTPsec project and the license was
intended to use the default project license for new code, BSD-2-Clause. The
notices regarding BSD-4-Clause-UC were copy-paste typos.

See https://lists.ntpsec.org/pipermail/devel/2020-April/009294.html for Hal's sign-off.

Additionally, we're removing the copyright for the University of California
from ntpd/ntp_dns.c. All code in this file was written by Hal as part of adding
NTS support to the project. The copyright assignment was a copy-paste typo.