Use a set instead of a list for python_scripts so we don't have duplicates
build() is called by both the "build" and "install" commands. So if
python_scripts is a list, then we actually insert each script into the list
twice and we end up attempting to install each script twice.  This is likely a
problem in the "pages" CI job failing periodically because we create some of
the install products more than once.