CI: shuffle asciidoc classic and py3 jobs to where they will run right ...
The job that was used to run the asciidoc py3 test got stuck installing
tex of all things and the jobs that were used to run asciidoc classic
tests had silently shifted to py3.

Move the classic jobs to Debian oldstable where the will run for a
while correctly and move the py3 job to AlpineLinux where it does
not have to worry about tex. Also, exploit our docker images for a
small speed boost.