ntpdig: Continue on errors in non-concurrent mode
In concurrent mode, select() errors were not fatal.  But in
non-concurrent mode, any error from select() would be fatal, even if
there was already a valid response received.

For example, I had a scenario where there was an A and AAAA returned
for a host (in that order).  The A worked but the AAAA did not.

This change makes the two code paths consistent.