Use latest docker image when building docker images
We were previously using the "stable" version of docker and this tag hasn't
been updated in 2 years. We instead switch to "latest" since we have no need to
pin to a specific version.

The "stable" version of docker was older than 20.10.10 and it caused issues for
linux distributions using glibc >= 2.34 because of how clone3() is used.

See https://github.com/AkihiroSuda/clone3-workaround for additional details.