ntp_control: fix kerninfo STA_NANO-sensitive and non-sensitive stats
Some of these stats are STA_NANO-sensitive ("KNUToMS") but some are always
in microseconds ("KUToMS") - treat them accordingly when converting to the
milliseconds that the NTP protocol uses.

In addition to the above, there was a typo in the code - some of the
variables had a "KToMS" flag, but the "v_kli" kernel variable handler only
tested for "ToMS" flag (a different one).