clockwork: move STA_NANO-related "constant" adjustment to ntp_adjtime_ns()
We are already adjusting other STA_NANO-related parameters here, adjust
"constant" here too.

Looking at the relevant kernel code it is also clear that "constant" cannot
go below zero only after having 4 added in the microseconds mode, so we
need to allow the input parameter to go as low as -4 in this mode.
Otherwise initial setting of zero "constant" will set it to 4 instead.