ntp_control.c: Fix a compiler warning
The expression was shifting a char left before promoting it to
a short.  This should actually be OK, since intermediate results
are supposed to be int or better, but it's more conceptually correct
to promote it first, which avoids a warning from GCC 4.2.  It's
also slightly better to do the mask before the shift.

TESTED:
Now builds without warnings and passes tests, with 28 compilers
from GCC 4.2 and Clang 3.3 to GCC 13 and Clang 17.