aes_siv: Remove troublesome -Wconversion
This warning is known to be overzealous in GCC 4.2, and apparently
toned down in later GCC versions.

According to devel/CommitLog-4.1.0, it was removed from the "classic"
CFLAGS in 2001-04-21, and hadn't resurfaced until it was applied
here solely to aes_siv components in 468fc5a7c9.

TESTED:
Now builds without warnings and passes tests, with 28 compilers
from GCC 4.2 and Clang 3.3 to GCC 13 and Clang 17.