First cut at fixing lockclock mode
The old Mills code implemented lockclock mode with a configure/build
time switch.  We switched that to a runtime test: loop_data.lockclock

But that didn't get turned on early enough so ntpd's initialization
did things like try to restore the drift setting from the drift file
which would have trampled the code that was doing the real timekeeping.

This patch does:
1: Patch refclock_local to use ntp_gettime() rather than ntp_adjtime()
   (The answer is the return code, not anything in the struct.)

2: Patch ntp_config to turn on loop_data.lockclock while parsing
   the config file rather than waiting for somebody to open the
   local driver so it can do it.

That pair lets us put a break at ntp_adjtime since nobody should call it.

Note that ntp_adjtime has a messy API.  It is a combined reader and
writer depending on various bits in a field.