Second step on fixing lockclock mode
I think this is a bug in the Mills code.

The comment for local_clock in ntp_loopfilter.c says:

 * Return codes:
 * -1   update ignored: exceeds panic threshold
 * 0    update ignored: popcorn or exceeds step threshold
 * 1    clock was slewed
 * 2    clock was stepped

When in lockclock mode it was returning 0.  The caller was
doing nothing rather than switching sys_leap from
LEAP_NOTINSYNC to LEAP_NOWARNING.  While stuck in
LEAP_NOTINSYNC, clients would ignore our responses.

The comment also says:
 * If lockclock is on, the only thing this routine does is set the
 * sys_rootdisp variable equal to the peer dispersion.
But the code doesn't do that.